requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
coverage==7.6.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
PyJWT==2.9.0
sqlparse==0.5.1
